# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API v2",
        "contact": {
            "email": "ybier@esme-solutions.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of orders",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a order",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ben Order"
                                    },
                                    "status": {
                                        "description": "Status of the point. 1 or 0",
                                        "type": "string",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Show details of an order",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the order to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the order to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Ben Order"
                                            },
                                            "status": {
                                                "description": "Status of the order. 1 or 0",
                                                "type": "string",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "operationId": "303a7b797887b66c7790c4c187b54790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the order to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deleted"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "clientCredentials": {
                        "authorizationUrl": "http://powerbat-laravel/oauth/authorize",
                        "tokenUrl": "http://powerbat-laravel/oauth/token",
                        "refreshUrl": "http://powerbat-laravel/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}